requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers>=3.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
numpy==1.24.3
torch>=2.0.0
transformers>=4.21.0
httpx==0.25.2
python-dotenv==1.0.0

# Text chunking dependencies
kss>=4.3.2
nltk>=3.9
tiktoken>=0.8.0

# Performance testing dependencies
openai>=1.0.0
scikit-learn>=1.3.0

# Document conversion dependencies
marker-pdf[full]>=0.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
